requirements.txt
torch==2.0.1+cu117
torchvision==0.15.2+cu117
torchaudio==2.0.2
numpy==1.*
webdataset>=0.2.5
regex
ftfy
tqdm
pandas
braceexpand
huggingface_hub==0.25.0
transformers[sentencepiece]==4.35.2
timm==0.9.7
fsspec

# additional
albumentations
pytorch_lightning==2.3.1
opencv_python==4.10.0.84
scikit_learn==1.5.1
segmentation_models_pytorch==0.3.4
pydicom
matplotlib
torchmetrics==1.4.0.*
open-clip-torch==2.26.1
openai==1.51.2
openai-clip==1.0.1
grad-cam==1.5.4
jupyter
jupyterlab
tensorboard
wandb
pydantic==2.10.6
--extra-index-url https://download.pytorch.org/whl/cu117